feat(java): :sparkles: added ios support
