+++ core-java/src/main/java/com/github/wasiqb/boyka/config/ui/mobile/device/ApplicationSetting.java
@@ -29,6 +29,7 @@ import lombok.Data;
  */
 @Data
 public class ApplicationSetting {
+    private String          bundleId;
     private boolean         external;
     private int             installTimeout = 30;
     private String          path;
