+++ core-java/src/main/java/com/github/wasiqb/boyka/enums/OS.java
@@ -26,5 +26,9 @@ public enum OS {
     /**
      * Android platform.
      */
-    ANDROID
+    ANDROID,
+    /**
+     * iOS platform.
+     */
+    IOS
 }
