+++ core-java/src/main/java/com/github/wasiqb/boyka/enums/PlatformType.java
@@ -32,7 +32,12 @@ public enum PlatformType {
      */
     API,
     /**
+     * iOS application.
+     */
+    IOS,
+    /**
      * Web application.
      */
     WEB
+
 }
