+++ core-java/src/test/resources/boyka-config.json
@@ -176,6 +176,7 @@
             "install_timeout": 180
           },
           "avd": {
+            "name": "Pixel_6_Pro",
             "headless": true
           }
         }
@@ -210,6 +211,7 @@
             "install_timeout": 180
           },
           "avd": {
+            "name": "Pixel_6_Pro",
             "headless": true
           }
         }
@@ -282,6 +284,36 @@
             "autoAcceptAlerts": true,
             "isRealMobile": true,
             "w3c": true
+          }
+        }
+      },
+      "test_local_sauce_ios": {
+        "server": {
+          "protocol": "HTTP",
+          "host": "127.0.0.1",
+          "port": 4724,
+          "base_path": "/wd/hub",
+          "session_override": true,
+          "allow_insecure": [
+            "get_server_logs"
+          ]
+        },
+        "device": {
+          "os": "IOS",
+          "version": "15",
+          "name": "iPhone 14 Pro",
+          "automation": "UI_AUTOMATOR",
+          "type": "VIRTUAL",
+          "server_install_timeout": 60,
+          "server_launch_timeout": 60,
+          "swipe": {
+            "distance": 25,
+            "max_swipe_until_found": 5
+          },
+          "application": {
+            "path": "/apps/ios/saucedemo.zip",
+            "wait_activity": "com.swaglabsmobileapp.MainActivity",
+            "install_timeout": 180
           }
         }
       }
