+++ package.json
@@ -30,25 +30,25 @@
   "packageManager": "yarn@3.3.1",
   "private": true,
   "devDependencies": {
-    "@commitlint/cli": "^17.4.1",
-    "@commitlint/config-conventional": "^17.4.0",
+    "@commitlint/cli": "^17.4.2",
+    "@commitlint/config-conventional": "^17.4.2",
     "@types/node": "^18.11.18",
     "@typescript-eslint/eslint-plugin": "^5.48.1",
     "@typescript-eslint/parser": "^5.48.1",
-    "commitlint": "^17.4.1",
+    "commitlint": "^17.4.2",
     "eslint": "^8.31.0",
     "eslint-config-google": "^0.14.0",
     "eslint-config-prettier": "^8.6.0",
-    "eslint-import-resolver-typescript": "^3.5.2",
-    "eslint-plugin-import": "^2.26.0",
+    "eslint-import-resolver-typescript": "^3.5.3",
+    "eslint-plugin-import": "^2.27.4",
     "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-react": "^7.31.11",
+    "eslint-plugin-react": "^7.32.0",
     "husky": "^8.0.3",
-    "lerna": "^6.4.0",
+    "lerna": "^6.4.1",
     "lerna-changelog": "^2.2.0",
     "lint-staged": "^13.1.0",
-    "nx": "^15.4.5",
-    "prettier": "^2.8.2",
+    "nx": "^15.5.1",
+    "prettier": "^2.8.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "ts-node": "^10.9.1",
